Fix warnings

+++ libcaf_net/test/pipe_socket.cpp
@@ -39,8 +39,10 @@ CAF_TEST(send and receive) {
   pipe_socket rd_sock;
   pipe_socket wr_sock;
   std::tie(rd_sock, wr_sock) = unbox(make_pipe());
-  CAF_CHECK_EQUAL(write(wr_sock, send_buf), send_buf.size());
-  CAF_CHECK_EQUAL(read(rd_sock, receive_buf), send_buf.size());
+  CAF_CHECK_EQUAL(static_cast<size_t>(write(wr_sock, send_buf)),
+                  send_buf.size());
+  CAF_CHECK_EQUAL(static_cast<size_t>(read(rd_sock, receive_buf)),
+                  send_buf.size());
   CAF_CHECK(std::equal(send_buf.begin(), send_buf.end(), receive_buf.begin()));
 }
